+++ src/actor/response.rs
@@ -1,0 +1,26 @@
+// Copyright 2020 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::Serialize;
+
+use crate::actor::{message_type::MessageType, response_type::ResponseType};
+
+/// The actor response type.
+#[derive(Serialize, Debug)]
+pub struct Response {
+    #[serde(flatten)]
+    response: ResponseType,
+    action: MessageType,
+}
+
+impl Response {
+    /// Creates a new response.
+    pub fn new(action: MessageType, response: ResponseType) -> Self {
+        Self { response, action }
+    }
+
+    /// The response's type.
+    pub fn response(&self) -> &ResponseType {
+        &self.response
+    }
+}
